requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
vonage
email-validator
google-generativeai
